requirements.txt
absl-py==0.9.0
aioredis==1.3.1
alabaster==0.7.12
asgiref==2.3.2
astor==0.8.1
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.1.3
Automat==0.8.0
Babel==2.8.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
channels-redis==2.3.2
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
Cython==0.29.14
daphne==2.4.1
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.10
docutils==0.16
entrypoints==0.3
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
imageio==2.6.1
imagesize==1.2.0
imgaug==0.3.0
importlib-metadata==1.5.0
incremental==17.5.0
ipykernel==5.1.4
ipyparallel==6.2.4
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
msgpack==0.5.6
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nose==1.3.7
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
opencv-python-headless==4.2.0.32
opt-einsum==3.1.0
packaging==20.1
pandocfilters==1.4.2
parso==0.6.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pycparser==2.19
Pygments==2.5.2
PyHamcrest==2.0.0
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
Rx==3.0.1
scikit-image==0.16.2
scipy==1.4.1
Send2Trash==1.5.0
Shapely==1.7.0
six==1.14.0
snowballstemmer==2.0.0
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==2.1.0
zope.interface==4.7.1
